celery_workers/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchaudio==2.2.0+cpu
torchvision==0.17.0+cpu
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
deprecation==2.1.0
distro==1.9.0
filelock==3.9.0
fsspec==2023.12.2
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.2
kombu==5.3.5
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
openai==1.10.0
packaging==23.2
pillow==10.2.0
pinecone-client==3.0.2
postgrest==0.15.0
prompt-toolkit==3.0.43
pydantic==2.6.0
pydantic_core==2.16.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
realtime==1.0.2
regex==2023.12.25
requests==2.28.1
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
sympy==1.12
tokenizers==0.15.1
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.13
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
google-generativeai==0.3.2
redis==5.0.1